requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.4
Click==7.0
colorama==0.4.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
mysql-connector==2.1.6
passlib==1.7.1
pluggy==0.8.0
py==1.7.0
pytest==3.9.1
python-dateutil==2.7.3
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
